Extract test doubles and passthrough coordination
